requirements.txt
torch>=2.2.0
transformers>=4.43.0
accelerate>=0.30.0
bitsandbytes>=0.43.0
langgraph>=0.1.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
tiktoken>=0.7.0
pymupdf>=1.24.0
python-docx>=1.1.0
ragas>=0.1.0
scikit-learn>=1.5.0
jinja2>=3.1.0
schedule>=1.2.0
python-dotenv>=1.0.0
numpy>=1.26.0
markdown>=3.6
xhtml2pdf>=0.2.11
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9
